Moves User::validateUsername() to Security::validateUsername()
